refactor: remove old widget hooks and old WidgetFrame
